refacto(pdb.acquisition): extract exception from PDBServer
